# sdk/trace: use sync.Pool in randomIDGenerator instead of Mutex