randomIDGenerator used a sync.Mutex to coordinate generation of TraceID
and SpanID. When used in relatively hot code paths this would cause
significant mutex contention as observed in pprof/mutex profiles.

This uses a sync.Pool instead which allows using generating and using
of *rand.Rand as needed without the need for a Lock.

```
name                 old time/op    new time/op    delta
SpanIDGeneration-10     191ns ± 5%      64ns ± 2%  -66.60%  (p=0.000 n=9+10)

name                 old alloc/op   new alloc/op   delta
SpanIDGeneration-10     0.00B          0.00B          ~     (all equal)

name                 old allocs/op  new allocs/op  delta
SpanIDGeneration-10      0.00           0.00          ~     (all equal)
```

Signed-off-by: Maisem Ali <maisem@tailscale.com>