@dmathieu 2022-09-01
Could you add a benchmark for this? It would then be nice to run it on the main branch, and with your changes and paste the results within this PR.
You can find examples of other benchmarks in the same package here: https://github.com/open-telemetry/opentelemetry-go/blob/main/sdk/trace/benchmark_test.go