@maisem 2022-09-01
@dmathieu I've added a benchmark. PTAL