@maisem 2022-09-01
Minor update to fix the lint check.