@MadVikingGod 2022-09-01
After Reruning with similar results, I think my question still stands.  To progress this you either need to rework it so that the pool performs equal or better under normal benchmarks, or there needs some compelling reason to accept this.

```
benchstat bench-main bench-3134 
name                                          old time/op    new time/op    delta
SpanProcessor-4                                 13.2µs ± 6%    13.7µs ± 3%  +3.49%  (p=0.019 n=10+10)
SpanProcessorVerboseLogging-4                   14.1µs ± 9%    14.4µs ± 7%    ~     (p=0.256 n=10+10)
SpanLimits/AttributeValueLengthLimit-4          10.8µs ± 7%    10.8µs ± 2%    ~     (p=0.214 n=9+9)
SpanLimits/AttributeCountLimit-4                9.40µs ± 5%    9.58µs ± 6%    ~     (p=0.280 n=10+10)
SpanLimits/EventCountLimit-4                    10.3µs ± 9%    10.3µs ± 4%    ~     (p=0.743 n=9+8)
SpanLimits/LinkCountLimit-4                     10.4µs ± 8%    10.7µs ± 9%    ~     (p=0.190 n=10+10)
SpanLimits/AttributePerEventCountLimit-4        10.7µs ± 8%    10.9µs ± 4%    ~     (p=0.113 n=10+9)
SpanLimits/AttributePerLinkCountLimit-4         10.9µs ± 7%    10.8µs ± 4%    ~     (p=0.661 n=9+10)
SpanSetAttributesOverCapacity-4                 2.25µs ± 6%    2.34µs ± 4%  +3.79%  (p=0.003 n=10+10)
StartEndSpan/AlwaysSample-4                      691ns ± 7%     711ns ± 3%    ~     (p=0.156 n=10+9)
StartEndSpan/NeverSample-4                       320ns ± 4%     346ns ± 6%  +8.30%  (p=0.000 n=10+10)
SpanWithAttributes_4/AlwaysSample-4             1.42µs ± 7%    1.40µs ± 2%    ~     (p=0.400 n=10+9)
SpanWithAttributes_4/NeverSample-4               505ns ± 6%     537ns ± 9%  +6.51%  (p=0.004 n=10+10)
SpanWithAttributes_8/AlwaysSample-4             1.96µs ±10%    1.97µs ± 2%    ~     (p=0.169 n=9+9)
SpanWithAttributes_8/NeverSample-4               672ns ± 6%     704ns ± 5%  +4.80%  (p=0.003 n=10+10)
SpanWithAttributes_all/AlwaysSample-4           1.80µs ± 8%    1.80µs ± 9%    ~     (p=0.987 n=10+10)
SpanWithAttributes_all/NeverSample-4             540ns ± 6%     563ns ± 5%  +4.34%  (p=0.004 n=9+10)
SpanWithAttributes_all_2x/AlwaysSample-4        2.70µs ± 5%    2.78µs ± 7%    ~     (p=0.139 n=9+10)
SpanWithAttributes_all_2x/NeverSample-4          761ns ± 4%     794ns ± 5%  +4.23%  (p=0.001 n=9+10)
SpanWithEvents_4/AlwaysSample-4                 1.70µs ± 9%    1.77µs ± 7%  +4.00%  (p=0.035 n=10+10)
SpanWithEvents_4/NeverSample-4                   329ns ± 3%     354ns ± 4%  +7.36%  (p=0.000 n=9+10)
SpanWithEvents_8/AlwaysSample-4                 2.59µs ± 7%    2.68µs ± 2%  +3.52%  (p=0.010 n=10+9)
SpanWithEvents_8/NeverSample-4                   342ns ± 4%     357ns ± 5%  +4.54%  (p=0.002 n=10+10)
SpanWithEvents_WithStackTrace/AlwaysSample-4    1.03µs ±10%    1.06µs ± 4%    ~     (p=0.050 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-4      369ns ± 4%     398ns ± 6%  +7.86%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/AlwaysSample-4     1.02µs ± 5%    1.05µs ± 5%  +2.53%  (p=0.040 n=9+9)
SpanWithEvents_WithTimestamp/NeverSample-4       418ns ± 7%     441ns ± 3%  +5.48%  (p=0.002 n=10+9)
TraceID_DotString-4                             88.7ns ± 2%    90.4ns ± 5%    ~     (p=0.065 n=9+10)
SpanID_DotString-4                              62.8ns ± 3%    64.6ns ± 2%  +2.81%  (p=0.003 n=9+9)

name                                          old alloc/op   new alloc/op   delta
SpanProcessor-4                                 8.75kB ± 0%    8.77kB ± 0%  +0.21%  (p=0.000 n=10+10)
SpanProcessorVerboseLogging-4                   9.91kB ± 0%    9.93kB ± 0%  +0.13%  (p=0.000 n=10+8)
SpanLimits/AttributeValueLengthLimit-4          10.6kB ± 0%    10.6kB ± 0%  +0.21%  (p=0.000 n=10+10)
SpanLimits/AttributeCountLimit-4                8.58kB ± 0%    8.59kB ± 0%  +0.20%  (p=0.000 n=10+10)
SpanLimits/EventCountLimit-4                    10.3kB ± 0%    10.3kB ± 0%  +0.20%  (p=0.000 n=10+10)
SpanLimits/LinkCountLimit-4                     10.3kB ± 0%    10.3kB ± 0%  +0.20%  (p=0.000 n=10+10)
SpanLimits/AttributePerEventCountLimit-4        10.5kB ± 0%    10.5kB ± 0%  +0.20%  (p=0.000 n=10+10)
SpanLimits/AttributePerLinkCountLimit-4         10.5kB ± 0%    10.5kB ± 0%  +0.20%  (p=0.000 n=10+10)
SpanSetAttributesOverCapacity-4                   496B ± 0%      496B ± 0%    ~     (all equal)
StartEndSpan/AlwaysSample-4                       432B ± 0%      432B ± 0%    ~     (all equal)
StartEndSpan/NeverSample-4                        128B ± 0%      128B ± 0%    ~     (all equal)
SpanWithAttributes_4/AlwaysSample-4             1.14kB ± 0%    1.14kB ± 0%  +0.18%  (p=0.000 n=10+8)
SpanWithAttributes_4/NeverSample-4                384B ± 0%      384B ± 0%    ~     (all equal)
SpanWithAttributes_8/AlwaysSample-4             1.90kB ± 0%    1.91kB ± 0%  +0.16%  (p=0.000 n=10+10)
SpanWithAttributes_8/NeverSample-4                640B ± 0%      641B ± 0%  +0.16%  (p=0.000 n=10+9)
SpanWithAttributes_all/AlwaysSample-4           1.71kB ± 0%    1.72kB ± 0%  +0.18%  (p=0.000 n=10+8)
SpanWithAttributes_all/NeverSample-4              448B ± 0%      448B ± 0%    ~     (all equal)
SpanWithAttributes_all_2x/AlwaysSample-4        3.06kB ± 0%    3.06kB ± 0%  +0.17%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/NeverSample-4           768B ± 0%      769B ± 0%  +0.13%  (p=0.000 n=10+10)
SpanWithEvents_4/AlwaysSample-4                   864B ± 0%      865B ± 0%  +0.12%  (p=0.000 n=10+10)
SpanWithEvents_4/NeverSample-4                    128B ± 0%      128B ± 0%    ~     (all equal)
SpanWithEvents_8/AlwaysSample-4                 1.31kB ± 0%    1.31kB ± 0%  +0.15%  (p=0.000 n=10+10)
SpanWithEvents_8/NeverSample-4                    128B ± 0%      128B ± 0%    ~     (all equal)
SpanWithEvents_WithStackTrace/AlwaysSample-4      544B ± 0%      545B ± 0%  +0.11%  (p=0.011 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-4       144B ± 0%      144B ± 0%    ~     (all equal)
SpanWithEvents_WithTimestamp/AlwaysSample-4       568B ± 0%      569B ± 0%  +0.18%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/NeverSample-4        168B ± 0%      168B ± 0%    ~     (all equal)

<<< Rmoved allocs/op because they were all equal >>>
```