@maisem 2022-09-02
> Explore ways to make the base case better. I would suggest preallocating one or more RNGs into the pool. This could even turn into an optimization you might want to make if you make the size of the preallocation a const that you override with a build flag.

Preloading the pool has no impact. There is a potential simple improvement that can be made using `mu.TryLock` however that was added in Go 1.18 and this project has a minimal version requirement of Go 1.17. The complexity of the other approaches is not worth it IMO.

> Build a benchmark demonstrating that this change improves span creation performance in a wide range of cases. If it's worse in 1 thread, the same in 3, and better at 5 it's a lot easier to take a change than if you need 100 threads to see the effect.

Added more sub-benchmarks to demonstrate improvement across multiple concurrent worker counts.
There is admittedly a degradation when there is only one worker, but arguably in that case you are not running in a performance critical system to begin with.

For all other cases the improvement is substantial.

```
name                      old time/op  new time/op  delta
SpanIDGeneration/1-10     35.0ns ± 0%  41.2ns ± 0%  +17.83%  (p=0.000 n=9+10)
SpanIDGeneration/2-10     91.6ns ± 1%  46.3ns ± 3%  -49.40%  (p=0.000 n=9+10)
SpanIDGeneration/3-10      120ns ± 1%    51ns ±10%  -57.37%  (p=0.000 n=9+10)
SpanIDGeneration/5-10      134ns ± 0%    66ns ± 0%  -50.57%  (p=0.000 n=9+9)
SpanIDGeneration/10-10     157ns ± 1%    64ns ± 1%  -59.51%  (p=0.000 n=10+10)
SpanIDGeneration/100-10    170ns ± 1%    64ns ± 1%  -62.49%  (p=0.000 n=10+10)
SpanIDGeneration/1000-10   172ns ± 1%    64ns ± 1%  -63.08%  (p=0.000 n=10+10)
```